change more names and redistribute main and main_utils into modules with more descriptive names, #45, #46
