+++ bacpipe/tests/test_embedding_creation.py
@@ -6,8 +6,11 @@ from pathlib import Path
 
 import bacpipe
 from bacpipe import EMBEDDING_DIMENSIONS
-from bacpipe.main_utils import run_pipeline_for_model, embeds_array_without_noise
-from bacpipe.main import Loader, Embedder
+from bacpipe.core.workflows import run_pipeline_for_model, embeds_array_without_noise
+
+from bacpipe.core.experiment_manager import Loader
+from bacpipe.model_pipelines.runner import Embedder
+
 from bacpipe.embedding_evaluation.label_embeddings import (
     generate_annotations_for_classification_task,
     make_set_paths_func,
